fix: corrected the button sizes add: added defualt values fix: renamed my data file
